+++ src/main/java/org/ecommerce/enums/Category.java
@@ -1,0 +1,5 @@
+package org.ecommerce.enums;
+
+public enum Category {
+    PARENT, CHILDREN
+}
