+++ src/main/java/org/ecommerce/services/impl/PasswordServiceImpl.java
@@ -1,8 +1,8 @@
 package org.ecommerce.services.impl;
 
+import org.ecommerce.enums.Error;
 import org.ecommerce.repositories.UserRepository;
 import org.ecommerce.services.PasswordService;
-import org.ecommerce.util.Error;
 
 import java.util.HashMap;
 import java.util.Map;
