+++ src/main/java/org/ecommerce/services/impl/UserServiceImpl.java
@@ -1,12 +1,12 @@
 package org.ecommerce.services.impl;
 
+import org.ecommerce.enums.Error;
 import org.ecommerce.models.Customer;
 import org.ecommerce.models.User;
 import org.ecommerce.repositories.UserRepository;
 import org.ecommerce.services.EmailService;
 import org.ecommerce.services.PasswordService;
 import org.ecommerce.services.UserService;
-import org.ecommerce.util.Error;
 
 import java.util.HashMap;
 import java.util.List;
