--- src/main/java/org/ecommerce/util/CategoryType.java
@@ -1,5 +1,0 @@
-package org.ecommerce.util;
-
-public enum CategoryType {
-    PARENT, CHILDREN
-}
