+++ src/main/java/org/ecommerce/util/Validator.java
@@ -1,0 +1,4 @@
+package org.ecommerce.util;
+
+public class Validator {
+}
